Fix SAML ticket verification failing when response is not namespaced
Namespacing of the soap response is not necessary. If the response
is not namespaced, the status code would be eg. `"Success"` instead
of `"ns2:Success"` and cause the client to fail authentication when
it had succeeded.

We can remove the `:` in the expected status code end substring
to support non namespaced responses.